+++ 223C/assignment5/find.c
@@ -4,4 +4,4 @@
 #include "student.h"
 
 extern struct Student* get(FILE*, long long);
-extern int search(FILE*, unsigned int);
+extern long long search(FILE*, unsigned int);
